issues/fs: Use sourcegraph/platform/storage instead of raw disk IO.
This is going to be the long term storage solution for most Sourcegraph
apps. Start using it in order to test platform/storage API, etc.

It's better because when the user removes a repo, it will be possible
to clean up app storage data for it in a reliable automatic way. It's
not possible to do that if apps manage their own per-repo storage.

/cc @slimsag