issues: Avoid false events; return events from Service.Edit method.
When an edit request results in no change state, do not emit an event.

This fixes an issue where closing an issue multiple times in a row would
create multiple "closed issue" events in a row.

Return events from Edit method. This de-duplicates the event logic,
which had to change and become slightly more complex, so that it's
present in only one place.