githubapi: Don't call IsCollaborator if error.
For example, if you don't have the auth scope to be able to check for
collaborators, then it'll return an error. Don't keep trying again for
each comment.

This functionally works at reducing unneeded API calls, but the code
style is probably not ideal. Need to think how to best solve this.