githubapi: Simplify detection of pull request vs issue.
GitHub API documentation says:

> If an issue is a pull request, the object will include a pull_request
> key.

Source: https://developer.github.com/v3/issues/#get-a-single-issue

So, it should be sufficient to check if PullRequestLinks is not nil.
Checking that PullRequestLinks.URL was not nil was only needed because
of a temporary bug on the GitHub API side. That bug has since been
fixed, see https://github.com/google/go-github/pull/357#r64128908.