fs: Fix List, Count error when issue tracker is empty.
Previously, a not exist error was being returned when an issue tracker
was completely empty. Now, it returns no error.

This is better, because issue trackers are often empty when new
repositories/packages are created. They become non-empty only after the
first issue is created.

This is similar to notifications/fs implementation.

Fixes:

	open /store/issues/repo/issues: no such file or directory