Add optional ListTimeline method; implement it in githubapi.
Clients of issues.Service generally need to call both ListComments and
ListEvents at once. It can be significantly more efficient to combine
them into one method that returns all results at once.

This is the case for the githubapi implementation that uses GraphQL.
Implement it there.

It's also simpler to have one method like this, so consider eventually
replacing ListComments and ListEvents with ListTimeline in the future
(as long as it works well with all clients).

Include IsTimelineLister method to enable wrapper implementations to
back out of implementing the optional TimelineLister interface.