fs: Use fromClose to populate event.Close field.
issues.Event.Close field is documented to be specified for Closed
events, so actually do populate it for Closed events.

This change is a no-op because the encoded closeDisk representation
of issues.Close{Closer: nil} happens to be nil *closeDisk, which is
the zero value.

But we should be doing this in case it ever becomes something else,
and so that fromClose isn't detected as unused.

fromClose will come in more handy when there are ways to close issues
via commits or changes.

Updates 53080de7ee7988534ed789de855ef1d2561ed80b.